+++ resources/ReleaseNotes.md
@@ -3,3 +3,4 @@
 - Make dependence on Parallel Computing Toolbox optional
 - Convert addition of missing row at end of file data into a processing step (`mag.process.Separate`)
 - Add `mag.process.Separate` to HK processing
+- Add export of whole science data